feat: updated nuxt documentation
This PR includes an application page for Nuxt within the documentation. Nuxt is becoming more popular as we go and I would love to add up to date documentation for fellow community members that are new to Coolify.

The page includes up to date steps to deploy the static Nuxt builds, the server Nuxt builds and nitro server builds. Running Nitro-only is like running Nuxt headless or without the frontend. Nitro, standalone, is best compared to Express.js.

Complementairy to this PR I have included another PR for coolify-examples where I included a healthcheck API endpoint and included an entire new Nitro standalone server build for the community to play with.